Extract load_entity_maps_pkl to utils
