+++ src/analyze.py
@@ -1,9 +1,7 @@
 from argparse import ArgumentParser
 from typing import Type, Mapping, Dict, List, Any
-import pickle
-import os
 
-from src.misc.settings import DataDir, DefaultFile
+from src.misc.utils import load_entity_maps_pkl
 from src.analyses.analysis import Analysis
 from src.analyses.basic_stats import BasicStatsAnalysis
 from src.analyses.duplicate_pipelines import DuplicatePipelinesAnalysis
@@ -49,13 +47,6 @@ def get_parser() -> ArgumentParser:
     return parser
 
 
-def load_entity_maps_pkl() -> Dict[str, dict]:
-    read_path = os.path.join(DataDir.EXTRACTION.value, DefaultFile.EXTRACTION_PKL.value)
-    print(f"Now loading pickled entity maps from '{read_path}'...")
-    with open(read_path, "rb") as f:
-        return pickle.load(f)
-
-
 def load_aggregations(req_aggs: List[str]) -> Dict[str, Any]:
     aggs = {}
     if len(req_aggs) == 0:
@@ -74,10 +65,9 @@ if __name__ == "__main__":
     parser = get_parser()
     args = parser.parse_args()
     entity_maps = load_entity_maps_pkl()
-    aggregations = load_aggregations(analysis.required_aggregations)
-
     analysis_class: Type[Analysis] = analysis_map[args.analysis]
     analysis = analysis_class()
+    aggregations = load_aggregations(analysis.required_aggregations)
 
     print(f"Now running {args.analysis} analysis...")
     analysis.run(entity_maps=entity_maps, aggregations=aggregations, verbose=args.verbose, refresh=args.refresh)
